+++ changelog.md
@@ -1,3 +1,4 @@
 - Fix mechanical mixers not finding GT mixing recipes without a circuit
 - Fix mechanical mixers running recipes when no stress is supplied
 - Fix improper logger warn when circuit is in recipe
+- Update ULV tier machine recipes